Add sorted() calls when listing directory
So that's the issue... Unix seems to read directory out of order
(usually reverse but sometimes random). While windows reads in order
everytime.

Nothing to do with the API - it called on season 1 and got 7 results but
was processing season 5 and needed 16 results. Coding at 1am..

Also the sorted() + inorder() call in test defeats the purpose?